Updated mustache files for client side validation
- Removed the cookies logic from the callApi method of the ApiClient
class generated from ApiClient.mustache
- Replaced the "<&returnType>" (which is replaced with
Campaign, Asset, ...) with "Object" in order to simply return the
server response in the calling method without deserializing the
response to a explicit class (like Campaign, Asset...) because the
deserializing method was calling the class constructor without
passing in the required parameters, thus making the code to throw
InvalidModelException. This change was made in the api.mustache file
- Updated the partial_model_generic.mustache file so that, when used
by the CLI it will use the model constraints from the OpenAPI spec
(if such constraints exist) and generate client side validation.
This validations wil be generated in the constructors of the model
classes.